refactor examples for c3 tests
