+++ examples/BUILD.bazel
@@ -1,0 +1,7 @@
+cc_binary(
+    name = "c3_test",
+    srcs = ["test/c3_test.cc"],
+    deps = [
+        ":c3",
+    ],
+)
